Add Beat type validation
