+++ pkg/apis/beat/v1beta1/webhook.go
@@ -14,27 +14,11 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
-
-	commonv1 "github.com/elastic/cloud-on-k8s/pkg/apis/common/v1"
-	"github.com/elastic/cloud-on-k8s/pkg/controller/common/version"
-	"github.com/elastic/cloud-on-k8s/pkg/utils/stringsutil"
 )
 
 var (
 	groupKind     = schema.GroupKind{Group: GroupVersion.Group, Kind: "Beat"}
 	validationLog = logf.Log.WithName("beat-v1beta1-validation")
-
-	defaultChecks = []func(*Beat) field.ErrorList{
-		checkNoUnknownFields,
-		checkNameLength,
-		checkSupportedVersion,
-		checkAtMostOneDeploymentOption,
-		checkImageIfTypeUnknown,
-	}
-
-	updateChecks = []func(old, curr *Beat) field.ErrorList{
-		checkNoDowngrade,
-	}
 )
 
 // +kubebuilder:webhook:path=/validate-beat-k8s-elastic-co-v1beta1-beat,mutating=false,failurePolicy=ignore,groups=beat.k8s.elastic.co,resources=beats,verbs=create;update,versions=v1beta1,name=elastic-beat-validation-v1beta1.k8s.elastic.co
@@ -91,45 +75,4 @@ func (b *Beat) validate(old *Beat) error {
 		return apierrors.NewInvalid(groupKind, b.Name, errors)
 	}
 	return nil
-}
-
-func checkNoUnknownFields(b *Beat) field.ErrorList {
-	return commonv1.NoUnknownFields(b, b.ObjectMeta)
-}
-
-func checkNameLength(ent *Beat) field.ErrorList {
-	return commonv1.CheckNameLength(ent)
-}
-
-func checkSupportedVersion(b *Beat) field.ErrorList {
-	return commonv1.CheckSupportedStackVersion(b.Spec.Version, version.SupportedBeatVersions)
-}
-
-func checkAtMostOneDeploymentOption(b *Beat) field.ErrorList {
-	if b.Spec.DaemonSet != nil && b.Spec.Deployment != nil {
-		msg := "Specify either daemonSet or deployment, not both"
-		return field.ErrorList{
-			field.Forbidden(field.NewPath("spec").Child("daemonSet"), msg),
-			field.Forbidden(field.NewPath("spec").Child("deployment"), msg),
-		}
-	}
-
-	return nil
-}
-
-func checkImageIfTypeUnknown(b *Beat) field.ErrorList {
-	knownTypes := []string{"filebeat", "metricbeat"}
-	if !stringsutil.StringInSlice(b.Spec.Type, knownTypes) &&
-		b.Spec.Image == "" {
-		return field.ErrorList{
-			field.Required(
-				field.NewPath("spec").Child("image"),
-				"Image is required if Beat type is not one of [filebeat, metricbeat]"),
-		}
-	}
-	return nil
-}
-
-func checkNoDowngrade(prev, curr *Beat) field.ErrorList {
-	return commonv1.CheckNoDowngrade(prev.Spec.Version, curr.Spec.Version)
 }
